+++ relevanceai/constants/messages.py
@@ -30,4 +30,8 @@ class Messages:
 
     TOKEN_MESSAGE = "Activation token (you can find it here: {} )\n"
 
+    INSERT_GOOD = "✅ All documents inserted/edited successfully."
+
+    INSERT_BAD = "❗Few errors with vectorizing documents. Please check logs."
+
     BUILD_HERE = "Build your clustering app here: "