+++ relevanceai/operations/base.py
@@ -1,7 +1,7 @@
 """
 Base class for operations.
 """
-from typing import Any, List, Optional
+from typing import Any, List, Union, Optional
 from relevanceai.client.helpers import (
     Credentials,
     process_token,
@@ -21,16 +21,16 @@ class BaseOps:
         return self(*args, **kwargs)
 
     @classmethod
-    def from_credentials(self, credentials: Credentials):
+    def from_credentials(self, credentials: Credentials, *args, **kw):
         return self(credentials=credentials)
 
     @classmethod
-    def from_token(self, token: str):
+    def from_token(self, token: str, *args, **kw):
         """
         If this is from a token, then we use this
         """
         credentials = process_token(token)
-        return self(credentials=credentials)
+        return self(credentials=credentials, *args, **kw)
 
     @classmethod
     def from_client(self, client, *args, **kwargs):
@@ -59,3 +59,11 @@ class BaseOps:
             *args,
             **kwargs,
         )
+
+    def _get_dataset_id(self, dataset: Union[str, Any]):
+        from relevanceai.dataset import Dataset
+
+        if isinstance(dataset, str):
+            return dataset
+        elif isinstance(dataset, Dataset):
+            return dataset.dataset_id