+++ relevanceai/operations/vector/__init__.py
@@ -1,2 +1,4 @@
 from relevanceai.operations.vector.vectorize import VectorizeOps
 from relevanceai.operations.vector.search import SearchOps
+from relevanceai.operations.vector.base import Base2Vec
+from relevanceai.operations.vector.vectorizer import Vectorizer